Move node metadata into metadata submodule
